- populated spinner in edituser - added navigation from user to edituser - added user fragment, viewmodel, and layout - added editUser method in FirestoreDatabase - change phone to be long in UsersDataModel

+++ app/src/main/java/com/example/first_responder_app/dataModels/UsersDataModel.java
@@ -12,12 +12,12 @@ public class UsersDataModel {
     private String first_name;
     private String last_name;
     private String password;
-    private int phone_number;
+    private Long phone_number;
     private String rank;
     private String username;
     private boolean is_responding;
 
-    public UsersDataModel(String address, String first_name, String last_name, String password, int phone_number, String rank, String username, boolean is_responding) {
+    public UsersDataModel(String address, String first_name, String last_name, String password, Long phone_number, String rank, String username, boolean is_responding) {
         this.address = address;
         this.first_name = first_name;
         this.last_name = last_name;
@@ -44,7 +44,7 @@ public class UsersDataModel {
 
     public String getPassword() { return password; }
 
-    public int getPhone_number() { return phone_number; }
+    public Long getPhone_number() { return phone_number; }
 
     public String getRank() { return rank; }
